Create Bank Account Assignment
This is a basic application which stores Bank Account objects, each of which has a Name and Balance. The user is offered choices via a menu, and the menu invokes functionality. Note: the Birthday method was an explicit request. It doesn't have much use otherwise, and was included as per instructions.